requirements.txt
annotated-types
anyio
blinker
certifi
cffi
cftime
charset-normalizer
click
contourpy
cryptography
cuda-bindings
cuda-pathfinder
cycler
decorator
faiss-cpu
filelock
Flask
fonttools
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h5py
hf-xet
httpcore
httplib2
httpx
huggingface_hub
idna
itsdangerous
Jinja2
joblib
kiwisolver
lxml
MarkupSafe
matplotlib
mpmath
netCDF4
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvshmem-cu12
nvidia-nvtx-cu12
obspy
packaging
pillow
playwright
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyee
pyparsing
python-dateutil
python-dotenv
PyYAML
regex
requests
rsa
safetensors
scikit-learn
scipy
segyio
sentence-transformers
setuptools
shellingham
six
SQLAlchemy
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
triton
typer-slim
typing-inspection
typing_extensions
uritemplate
urllib3
Werkzeug
